demo/requirements.txt
pyinstaller
python-dotenv~=1.0.1
langchain~=0.2.14
langgraph~=0.2.5
langchain_experimental
langchain-google-vertexai~=1.0.8
beautifulsoup4
pandas
faiss-cpu
chroma
chromadb==0.5.3
langchain-core~=0.2.33
google-cloud-aiplatform~=1.63.0
langchain-community~=0.2.12
langchain-text-splitters~=0.2.2